+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Converted ``Index has bad magic - bypassing index`` errors to warnings https://github.com/oleg68/GrandOrgue/issues/58
 # 0.3.1.2341-9.os (2021-08-07)
 - Fixed ``A device wants different samples per buffer: 15`` on OSx  with RtAudio https://github.com/oleg68/GrandOrgue/issues/54
 - Displaying actual error text instead of ``Internal Portaudio Error`` https://github.com/oleg68/GrandOrgue/issues/52